Split genesis creation into a separate class.
